hello

+++ src/Player.java
@@ -1,7 +1,7 @@
 import java.awt.*;
 
 public class Player extends Entity{
-    private int attackRadius = 30;
+    private double attackRadius = 30;
 
     public Player(int x, int y) {
         super("Player", 100, 25, 10, x, y, 10, 10, Color.BLACK);
@@ -11,7 +11,7 @@ public class Player extends Entity{
     public boolean isPlayer() {
         return true;
     }
-    public int getAttackRadius(){
+    public double getAttackRadius(){
         return attackRadius;
     }
 }
